+++ server/graphql/general/src/mutations/insert_insurance.rs
@@ -1,28 +1,14 @@
 use async_graphql::*;
 use graphql_core::standard_graphql_error::StandardGraphqlError;
 use graphql_core::{standard_graphql_error::validate_auth, ContextExt};
-use repository::name_insurance_join_row::{InsurancePolicyType, NameInsuranceJoinRow};
-use serde::Serialize;
+use graphql_types::types::IdResponse;
+use repository::name_insurance_join_row::NameInsuranceJoinRow;
 use service::{
     auth::{Resource, ResourceAccessRequest},
     insurance::insert::{InsertInsurance as ServiceInput, InsertInsuranceError as ServiceError},
 };
 
-#[derive(Enum, Copy, Clone, PartialEq, Eq, Debug, Serialize)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum InsurancePolicyNodeType {
-    Personal,
-    Business,
-}
-
-impl InsurancePolicyNodeType {
-    pub fn to_domain(&self) -> InsurancePolicyType {
-        match self {
-            InsurancePolicyNodeType::Personal => InsurancePolicyType::Personal,
-            InsurancePolicyNodeType::Business => InsurancePolicyType::Business,
-        }
-    }
-}
+use crate::types::InsurancePolicyNodeType as InsertInsurancePolicyNodeType;
 
 #[derive(InputObject)]
 pub struct InsertInsuranceInput {
@@ -31,7 +17,7 @@ pub struct InsertInsuranceInput {
     pub insurance_provider_id: String,
     pub policy_number_personal: String,
     pub policy_number_family: String,
-    pub policy_type: InsurancePolicyNodeType,
+    pub policy_type: InsertInsurancePolicyNodeType,
     pub discount_percentage: f64,
     pub expiry_date: chrono::NaiveDate,
     pub is_active: bool,
@@ -65,21 +51,10 @@ impl InsertInsuranceInput {
     }
 }
 
-pub struct InsertInsuranceNode {
-    pub id: String,
-}
-
-#[Object]
-impl InsertInsuranceNode {
-    pub async fn id(&self) -> &str {
-        &self.id
-    }
-}
-
 #[derive(Union)]
 #[graphql(name = "InsertInsuranceResponse")]
 pub enum InsertInsuranceResponse {
-    Response(InsertInsuranceNode),
+    Response(IdResponse),
 }
 
 pub fn insert_insurance(
@@ -109,9 +84,7 @@ pub fn map_response(
     from: Result<NameInsuranceJoinRow, ServiceError>,
 ) -> Result<InsertInsuranceResponse> {
     match from {
-        Ok(insurance) => Ok(InsertInsuranceResponse::Response(InsertInsuranceNode {
-            id: insurance.id,
-        })),
+        Ok(insurance) => Ok(InsertInsuranceResponse::Response(IdResponse(insurance.id))),
         Err(error) => map_error(error),
     }
 }
