+++ server/graphql/general/src/queries/insurances.rs
@@ -8,12 +8,11 @@ use graphql_core::{
     ContextExt,
 };
 use repository::name_insurance_join_row::{
-    InsurancePolicyType, NameInsuranceJoinRow, NameInsuranceJoinSort, NameInsuranceJoinSortField,
+    NameInsuranceJoinRow, NameInsuranceJoinSort, NameInsuranceJoinSortField,
 };
-use serde::Serialize;
 use service::auth::{Resource, ResourceAccessRequest};
 
-use crate::types::InsuranceProviderNode;
+use crate::types::{InsurancePolicyNodeType, InsuranceProviderNode};
 
 #[derive(Enum, Copy, Clone, PartialEq, Eq)]
 #[graphql(rename_items = "camelCase")]
@@ -29,23 +28,6 @@ pub struct InsuranceSortInput {
     /// Sort query result is sorted descending or ascending (if not provided the default is
     /// ascending)
     desc: Option<bool>,
-}
-
-#[derive(Enum, Copy, Clone, PartialEq, Eq, Debug, Serialize)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum InsurancePolicyNodeType {
-    Personal,
-    Business,
-}
-
-impl InsurancePolicyNodeType {
-    pub fn from_domain(policy_type: &InsurancePolicyType) -> InsurancePolicyNodeType {
-        use InsurancePolicyType::*;
-        match policy_type {
-            Personal => InsurancePolicyNodeType::Personal,
-            Business => InsurancePolicyNodeType::Business,
-        }
-    }
 }
 
 #[derive(PartialEq, Debug)]
