+++ server/graphql/general/src/types.rs
@@ -1,6 +1,7 @@
 use async_graphql::*;
 use chrono::NaiveDate;
-use repository::InsuranceProviderRow;
+use repository::{name_insurance_join_row::InsurancePolicyType, InsuranceProviderRow};
+use serde::Serialize;
 use service::{
     requisition_line::chart::{
         ConsumptionHistory, ItemChart, StockEvolution, SuggestedQuantityCalculation,
@@ -171,6 +172,30 @@ impl ItemChartNode {
             suggested_quantity_calculation: SuggestedQuantityCalculationNode {
                 suggested_quantity_calculation,
             },
+        }
+    }
+}
+
+#[derive(Enum, Copy, Clone, PartialEq, Eq, Debug, Serialize)]
+#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
+pub enum InsurancePolicyNodeType {
+    Personal,
+    Business,
+}
+
+impl InsurancePolicyNodeType {
+    pub fn from_domain(policy_type: &InsurancePolicyType) -> InsurancePolicyNodeType {
+        use InsurancePolicyType::*;
+        match policy_type {
+            Personal => InsurancePolicyNodeType::Personal,
+            Business => InsurancePolicyNodeType::Business,
+        }
+    }
+
+    pub fn to_domain(&self) -> InsurancePolicyType {
+        match self {
+            InsurancePolicyNodeType::Personal => InsurancePolicyType::Personal,
+            InsurancePolicyNodeType::Business => InsurancePolicyType::Business,
         }
     }
 }
